Refactor debuginfo to use feature flags
